05.docker/salary_counter/requirements.txt
asgiref==3.4.1
DAWG-Python==0.7.2
Django==3.2.7
django-ratelimit==3.0.1
djangorestframework==3.12.4
docopt==0.6.2
MarkupSafe==2.0.1
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pytz==2021.1
sqlparse==0.4.2
psycopg2-binary
gunicorn
django-environ
dj_database_url